Issue: ProxyBuilder.getMethodsToProxyRecursive() was not returning list of methods in deterministic order.
RCA: In order to determine which methods to proxy, ProxyBuilder populates hash sets of MethodSetEntry instances. This prevents duplicate entries for methods from occurring. Since the hash set is unordered and the resulting list returned by ProxyBuilder.getMethodsToProxyRecursive() is built from the hash set, the order of the list is non-deterministic.

Fix: After building the list, sort it by Method.toString(), which contains all of the information in a method's signature.