Update to Mockito 2.2.29
This commit also changes the next snapshot version to be
2.2.0 instead of 1.6.0. The intention is to make the major
and minor version of dexmaker stay in line with the major
and minor version of Mockito that is supported. Mockito's
patch version is updated with every commit released, and we
may want releases of dexmaker that don't change the version
of Mockito that we depend on so there's not much benefit in
staying in sync with that.

Also worth noting, this commit doesn't do anything to support
mocking final methods. This just fixes dexmaker to work with
Mockito 2.2.x core.