Allow mocking of final classes/methods on Android
by adding dexmaker-mockito-inline.

This is very similar to how mockito-inline works for JVMs but replaces
bytebuddy with a jvmti agent that uses the slicer lib.

If a class gets mocked for the first time the agent adds an entry hook
to all non-abstract methods. This entry hook calls into
MockMethodDispatcher->MockMethodAdvice which checks if the current
object is mocked. If so it executes
InvocationHandlerAdapter#doIntercept. The entryhooks are added as raw
DEX byte code in agent.cc.

Abstract methods are mocked via subclassing as in dexmaker-mockito. Of
course we don't have to mock non-abstract methods twice, hence we
restict the methods overridden by the subclasses.

This passes 99% of the mockito tests.
- Serialization is not supported yet
- Some mockito tests rely on bytebuddy and hence don't work on Android. We
  should abtract them out to be able to use dexmaker instead

I added some basic tests for mocking of final methods as
dexmaker-mockito-inline-tests. Further I fixed dexmaker-tests that were
failing on my current aosp-master build.