Multiple Jvmti agents can't interfere with mocking
If a class transformation is requested by any agent, the transformation
starts from the original code. Hence if we previosly added entry hooks
to a class, we now have to add them again. To do that we call into
ClassTransformer#transform every time a class get transformed, not only
when this was triggerd via ClassTransformer#mockClass.

Also added a test that simulates this behavior.