Check if we should transform before touching classBuffer
The class transformation hook is called every time a class is loaded or
an re-transformation is requested. Hence in normal java VM it is called
frequently as new classes are loaded. We usually don't want to transform
those newly loaded classed.

We cannot just unregister the hook when no mocking-transformtion is in
progress as other JVMTI agents might request a retransformation of an
already loaded class. This transformation will start with an completely
untransformed class - without the entry hook our agent adden. Hence in
this case our agent has to intercept it an load the entry-hooks again.

Touching the class file buffer is expensive as the JVM generates the
content on the first read access. Hence try to avoid it if not necessary.