Issue #64: Don't use internal mockito WeakConcurrentMap and ..Set
The MockMap is mostly a re-implementation of WeakConcurrentMap with a
few Android'y twists such as creating less objects and using
android.util types where they don't cause a performance regression.

I stopped using the WeakConcurrentSet as we need to remember which
classes are transformed, hence using a weak-reference would require us
to un-transform the classes when the reference goes away. We don't do
that, hence don't use weak-references.