Set up dexmaker-mockito-inline for publishing
Since this artifact publishes an aar file instead
of a jar, we can't reuse publishing.gradle directly.
We could later improve the publishing scripts so
there isn't so much duplication, but for now this
should be enough to get the new artifact published.

Tested this change by running :dexmaker-mockito-inline:publishToMavenLocal
and checking that the generated pom file included the
proper dependencies. Also created a sample Android app
and verified that I can depend on this artifact via
mavenLocal() and could successfully write a test that
mocks a final class / method.