Allow to stub static method via doReturn
The syntax for instance methods is

doReturn(toReturn).when(mock).instanceMethod(arguments);

The interesting part here is that when(mock) returns a object of the
mock's class. For static methods the equivalent would be

doReturn(toReturn).when(MockClass).staticMethod(arguments);

In this case it is not clear what 'when' should return. E.g. if we would
return an instance of MockClass we would effectvely calls
mockObject.staticMethod which would light all code checkers on fire as
you are not supposed to call static methods on objects.

Hence we use lambdas to set up stubbing, e.g.

doReturn(toReturn).when(() -> MockClass.staticMethod(arguments));

There are various problems with that

== lambdas can be more than a single method call ==

A lambda can be arbirary complex but we really only want a execute a single
method call to declare what should be stubbed.

Hence we need to check that there was only a single stubbable method
call and throw an error if not

== What class is the stubbing to be set up ==

When 'when' is called we don't know what class the stubbing should be
set up on. Hence we wait for any method interception and then set up
then stubbing just before we continue with the execution of the method.