Disable Gradle configureondemand - Fixes #101
Configuration on demand + the JFrog plugin seems to cause
Gradle sync issues in Android Studio. Disabling this option
and disabling the Configure on Demand setting in Android Studio
preferences fixes the issue.

This project doesn't need Configure on Demand anyway, so there's
no downside to disabling this option, even if Gradle and JFrog
come up with a different workaround later.

For more details, see:
https://github.com/JFrogDev/build-info/issues/86