Add spyOn(Object) that allows to turn existing object into a spy
This is meant to solve the problem that certain Android objects (e.g.
Activities) need to be created by the system to be useful.

While this is a feature of the ExtendedMockito it has a hook into
InlineDexmakerMockMaker. The other option would be to add a full inline
dexmaker mock maker only for this feature.