Issue #106: Allow mocks to call blacklisted APIs
As it is not clear what classes have blacklisted APIs any mock might be for a
such a class. Hence we have to mark any class as trusted to be able to
call the real methods.

Further the leniant copy tool has to be marked as trusted so that it can
copy state as some fields might be blacklusted too.