Do not reference abstract super methods
In API 19 and below generating code that calls abstract super methods
may cause verification errors.

This effect is only seen if the code was originally compiled against a
higher SDK level that had the method, but then the code was run against
a lower SDK level that does not have the method. This is a trick
androidx (aka. the android support library) is commonly using.

There is really no point of even generating the code that calls the
abstract super method, hence this change replaces the calling code by a
throw(new AbstractMethodError("'method' cannot be called').

The code referenced in ProxyBuilder#generateCodeForAllMethods was not
complete and making it complete would make it much more complicated.
Hence I removed it. If somebody wants to look at the actually
generated code, they can use 'dexdump -d' on the generated dex file.

Mockito handles attempts to call the real abstract method gracefully. It
does not even allow to set up such a stubbing.

Bonus:
- Added test for shared classloader functionality of the proxy builder
- I checked all way back to API19 and made sure that tests that cannot
  run on older API levels are ignored