Allow starting of static-spy-session for objects without 0-arg constructor
when calling Mockito.spy(Class c) mockito tries to create an object of
'c' using the 0-args constructor and fails if this is not possible.

For static spying the created object is just a marker and is not used
for anything, hence there is no need to create a correctly constructed
objects. Hence creating a mock is enough. We still have to call the
real-methods by default as this settings also applies to static methods.

I added a test that previously failed at spyStatic()