Allow for AccessFlags.ACC_SYNTHETIC.
AFAICT, there should be no reason this shouldn't be supported. It's
just adding an extra flag to the Method. Given that DexMaker is used
to generate code, it makes sense that we're going to frequently
generate synthetic methods.

I didn't really test this, except for validating that the constants in
Modifier, matches the constants in AccessFlags.